requirements.txt
django>=4.2,<6.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0

